+++ src/test/groovy/com/cedarsoftware/ncube/TestNCubeRuntime.groovy
@@ -3,6 +3,7 @@ package com.cedarsoftware.ncube
 import com.cedarsoftware.util.GuavaCache
 import groovy.transform.CompileStatic
 import org.junit.Test
+import org.springframework.cache.Cache
 
 import static com.cedarsoftware.ncube.NCubeAppContext.getNcubeRuntime
 import static com.cedarsoftware.util.TestUtil.assertContainsIgnoreCase
@@ -16,7 +17,7 @@ class TestNCubeRuntime extends NCubeBaseTest
     {
         populateCache()
         ncubeRuntime.clearCache(ApplicationID.testAppId, ['testbranch', 'testage'])
-        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).nativeCache.asMap()
+        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).asMap()
         assert 2 == cache.size()
         assert !cache.containsKey('testbranch')
         assert !cache.containsKey('testage')
@@ -31,10 +32,10 @@ class TestNCubeRuntime extends NCubeBaseTest
         assert ncubeRuntime.isCached(ApplicationID.testAppId, 'testage')
         assert ncubeRuntime.isCached(ApplicationID.testAppId, 'testcube')
         ncubeRuntime.clearCache(ApplicationID.testAppId)
-        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).nativeCache.asMap()
-        assert !cache.containsKey('testbranch')
-        assert !cache.containsKey('testage')
-        assert !cache.containsKey('testcube')
+        Cache cache = testClient.getCacheForApp(ApplicationID.testAppId)
+        assert !cache.get('testbranch')
+        assert !cache.get('testage')
+        assert !cache.get('testcube')
     }
 
     @Test
@@ -42,7 +43,7 @@ class TestNCubeRuntime extends NCubeBaseTest
     {
         populateCache()
         ncubeRuntime.clearCache(ApplicationID.testAppId, null)
-        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).nativeCache.asMap()
+        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).asMap()
         assert 0 == cache.size()
     }
 
@@ -77,7 +78,7 @@ class TestNCubeRuntime extends NCubeBaseTest
         createRuntimeCubeFromResource('test.branch.1.json')
         createRuntimeCubeFromResource('test.branch.age.1.json')
         createRuntimeCubeFromResource('testCube1.json')
-        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).nativeCache.asMap()
+        Map cache = ((GuavaCache) testClient.getCacheForApp(ApplicationID.testAppId)).asMap()
         return cache
     }
 }